moved filename parsers to images
